[CIR][Codegen] Fix bitfields unary and binary ops (#477)
This PR fixes a couple of  NIY features for bit fields. 

Basically, such expressions with bit fields `x->a++` and `x->a |= 42`
are supported now.

The main problem is `UnOp` verification - now it can be defined both by
`loadOp` and by `GetBitfieldOp` or even by `CastOp`. So shame on me, I
removed a test from `invalid.cir`