[CIR][Lowering] More cir.asm lowering (#472)
This PR adds lowering for `cir.asm`.

Also, two flags were added to the `cir.asm` : `hasSideEffects` and
`isStackAligned` in order to match with the llvm dialect.

Also, I added several simple tests for lowering.

I'm not sure but most likely the next PR will be the last one in this
story about assembly support )