[CIR] shufflevector and convertvector built-ins
Implement `__builtin_shufflevector` and `__builtin_convertvector` in
ClangIR.  This change contributes to the implemention of issue #284.

`__builtin_convertvector` is implemented as a cast.  LLVM IR uses the
same instructions for arithmetic conversions of both individual scalars
and entire vectors.  So ClangIR does the same.  The code for handling
conversions, in both CodeGen and Lowering, is cleaned up to correctly
handle vector types.  To simplify the lowering code and avoid `if
(type.isa<VectorType>())` statements everywhere, the utility function
`elementTypeIfVector` was added to `LowerToLLVM.cpp`.

`__builtin_shufflevector` has two forms, only one of which appears to be
documented.

The documented form, which takes a variable-sized list of integer
constants for the indices, is implemented with the new ClangIR operation
`cir.vec.shuffle.ints`.  This operation is lowered to the
`llvm.shufflevector` op.

The undocumented form, which gets the indices from a vector operand, is
implemented with the new ClangIR operation `cir.vec.shuffle.vec`.  LLVM
IR does not have an instruction for this, so it gets lowered to a long
series of `llvm.extractelement` and `llvm.insertelement` operations.