[CIR] Add `cir.dyn_cast` operation (#483)
This PR adds the `cir.dyn_cast` operation for representing
`dynamic_cast` in C++. It contains the following contents:

- [x] A new `cir.dyn_cast` operation.
- [x] ~Two new attributes that will be attached to `cir.dyn_cast`
operations:~
- [x] ~`#cir.dyn_cast_info` attributes, which gives general information
about a dynamic cast (e.g. the source RTTI pointer, the dest RTTI
pointer, etc.)~
- [x] ~`#cir.downcast_info` attribute, which gives even more detailed
information about a dynamic cast that is a down-cast. These information
will be used when rewriting the `cir.dyn_cast` operation with more
fundamental CIR operations.~
- [x] CIRGen support for the new operation and attributes.
- [x] Rewrite the new operation with more fundamental CIR operations in
LoweringPrepare.

~This is a draft PR. Now I only added the new operation / attributes,
and updated the CIRGen part. The LoweringPrepare for the new operation
is not implemented. Hopefully the draft can get some initial feedbacks
from the community and make sure it is on the right direction so we
don't waste time on wrong things.~

Related issue: #470 .