[CIR] Generate cir.dyn_cast for dynamic casts to void ptr
This patch update the CIRGen of `dynamic_cast` expressions and make it start to
generate `cir.dyn_cast` operations for `dynamic_cast` expressions that cast to
a void pointer. This patch also updates the lowering prepare pass so that it
lowers such `cir.dyn_cast` operations to the code emitted before this patch.