[CIR][CodeGen] Adds SwitchOp flattening (#549)
This PR adds flattening for `SwitchOp`. Despite of the previous PRs,
here we have to introduce an operation for the flattening, since later
we'll need to create `llvm.switch` in the lowering.
So `cir.flat.switch` is a new operation, which barely copied from the
dialect. I added several tests as well.