[CIR][Asm] Remove duplicated lambda & coroutine attributes (#580)
Do not print in cir.func definition the 'attr { ... }' with coroutine or
lambda attributes since they are already printed before the function
name. Otherwise redundancy breaks a future parsing. Sort the attributes
to be skipped so it is more obvious to see the list of attributes.
Improve the tests to check there are no spurious attributes anymore.