[CIR][Transforms] Simplify redundant bitcasts (#591)
Fix #479 .

There are three available stages to place the simplification in.

* A straightforward method is to extend `fold` method for CastOp.
But CIR does not use CanonicalizerPass, so it does not work.
* As for somehow equivalent to it, append a pattern to
`MergeCleanupsPass`.
But now it is mainly for CFG-related simplifications like block merging.
I don't know if this is the proper way. Shall we rename it to a broader
definition?
* Add a new pass for this issue.
This is definitely not very reasonable XD. We won't consider it unless
we're really out of options.

This PR includes the second option. What do you think @bcardosolopes ?