[CIR][CodeGen] Support trailing_zeros for constant string literals
The patch resolves [issue #248](https://github.com/llvm/clangir/issues/248). It can be considered a subsequent patch to [#373](https://github.com/llvm/clangir/pull/373), where the case of empty strings was processed.

The new patch adds processing for non-empty strings that may contain trailing zeros, such as:
```
char big_string[100000] = "123";
```
That is converted to
```
@big_string = #cir.const_array<"123" : !cir.array<!s8i x 3>, trailing_zeros> : !cir.array<!s8i x 100000>
```

It's worth noting that ordinary strings that are terminated with a single zero also get the trailing zeros, for instance
```
char string[] = "whatnow";
```
Is converted to
```
@string = #cir.const_array<"whatnow" : !cir.array<!s8i x 7>, trailing_zeros> : !cir.array<!s8i x 8>
```