[CIR][ABI][NFC] Add CC lowering for void CallOps
This patch implements the lowering of function calls that receive and
return void. In practice, nothing has to be done (at least for the x86
ABI), so this case is used as a primer for the target lowering library
since it helps populate the base logic for handling calling convention
lowering of function calls without adding too much complexity.