[CIR][Fix] FP builtins should lower directly to LLVM builtins
LLVM lowering for the following operations is introduced in #616 and #651: cos,
exp, exp2, log, log10, log2, sin, sqrt, fmod, and pow. However, they are not
lowered to their corresponding LLVM intrinsics; instead they are transformed to
libc calls during lowering prepare. This does not match the upstream behavior.

This patch tries to correct this mistake.