[CIR][ABI][NFC] Add CC lowering for void FuncOps (#678)
This patch implements the lowering of function definitions with no
arguments and returns. In pratice, nothing has to be done (at least for
the x86 ABI), so this case is used as a primer for the target lowering
library since it helps populate the base logic for handling calling
convention lowering of function definitions.