[CIR] Add FuncAttrs to cir.calls (#637)
Some function attributes are also callsite attributes, for instance,
nothrow. This means they are going to show up in both. We don't support
that just yet, hence the PR.

CIR has an attribute `ExtraFuncAttr` that we current use as part of
`FuncOp`, see CIROps.td. This attribute also needs to be added to
`CallOp` and `TryCalOp`.

Right now, In `CIRGenCall.cpp: AddAttributesFromFunctionProtoType` fills
in `FuncAttrs`, but doesn't use it for anything. We should use the
`FuncAttrs` result to populate constructing a `ExtraFuncAttr` and add it
to the aforementioned call operations.