[CIR][CIRGen] Add support for exact dynamic cast
This patch is the last piece to make CIR catch up upstream CodeGen on
dynamic_cast support. It ports an upstream optimization "exact cast" to CIR.

The basic idea of exact cast is when dynamic_cast to a final class, we don't
have to call into the runtime -- we could just check if the dynamic type of the
source object is exactly the destination type by quickly comparing the vtable
pointers.

This patch also fixes a bug in existing dynamic_cast CIRGen code. The bug
mistakenly removes the insertion point after emitting a call to bad_cast,
causing the CIRGen of any follow up statements to crash.