[CIR][ABI] Add X86_64 and AArch64 bool CC lowering
Implements calling convention lowering of bool arguments and return
value calling conventions for X86_64 and AArch64.

For x86_64, this is a bit of an odd case. In the orignal codegen bools
are represented as i8 everywhere, except for function arguments/return
values. In CIR, we don't allow i1 types, so bools are still represented
as `cir.bool` when in functions. However, when lowering to LLVM Dialect,
we need to ensure bools will be converted to i1 when in function's
argument/return values.