[CIR][Transforms][NFC] Use `unique_ptr` to encapsulate LowerModule (#752)
Currently `LowerModule` mimics `CodeGenModule` and uses many raw
references. It cannot be moved or copied. Value semantic does not fit
the need. For example, we cannot pass LowerModule around.

A better practice would be to use `unique_ptr` to encapsulate it. In the
future, we hold its ownership in some long-lived contexts (it's
`CodeGeneratorImpl` for `CodeGenModule`) and pass references to it
around safely.