[CIR][CodeGen][Lowering] Better handling of alloca address space with unified AS (#682)
`TargetCodeGenInfo::getASTAllocaAddressSpace` is a bad design because it
requires the targets return `LangAS`, which enforce the targets to
consider which languages could be their upstream and unnecessarily
complicate the target info. Unified AS in CIR is a better abstraction
level for this kind of target info.

Apart from that, the languages also has some requirements on the result
address space of alloca.

```cpp
void func() {
  int x;
  // Here, the AS of pointer `&x` is the alloca AS defined by the language.
}
```

When we have inconsistency between the alloca AS defined by the language
and the one from target info, we have to perform `addrspacecast` from
the target one to the language one.

This PR includes
* New method `CGM.getLangTempAllocaAddressSpace` which explicitly
specifies the alloca address space defined by languages. It replaces the
vague `LangAS::Default` in the AS comparisons from OG CodeGen.
* Replace `getASTAllocaAddressSpace` with `getCIRAllocaAddressSpace`,
which returns CIR unified AS.
* Also use `getCIRAllocaAddressSpace` as the result address space of
`buildAlloca`.
* Fix the lowering of `cir.alloca` operation to be address-space-aware.

We don't perform any `addrspacecast` in this PR, i.e. all the related
code paths still remain NYI and it's fine. That's because we don't even
have a `(language, target)` pair holding the inconsistency.

After these changes, in the previous OpenCL testcases we will see all
the alloca pointers turning into private AS, without any
`addrspacecast`.