[CIR][LowerToLLVM][NFC] Add data layout verification of alloca as (#769)
There are two sources for the target allocation address space: one from
`TargetCIRGenInfo::getCIRAllocaAddressSpace()` and another from
`targetDataLayout.getAllocaMemorySpace()`. Since both are provided by
the specific target, they should be consistent. This PR adds a check to
ensure this consistency and avoid potential errors.

The ctor of `CIRAllocaLowering` pattern is updated to pass the data
layout in.