[CIR][Dialect] Add address space attribute to global op (#779)
This PR adds the CIR address space attribute to GlobalOp and starts to
resolve the missing feature flag `addressSpaceInGlobalVar`.

The same asm format in pointer type is used:

```
cir.global external addrspace(offload_global) @addrspace1 = #cir.int<1> : !s32i
```

The parsing and printing helper is extracted into a common function to
be reused by both `GlobalOp` and `PointerType` with two custom format
proxy to it. That's because the signature of ODS generated method
differs from the one for PointerType.

Lowering to LLVM IR and CIRGen will come sequentially.