[CIR][CodeGen][Dialect][NFC] Refactor OpenCL stuff to separate impl files (#781)
This PR refactors OpenCL-specific logic that is rather independent from
CIRGen and dialect definition into two new implementation files:
`CIRGenOpenCL.cpp` and `CIROpenCLAttrs.cpp`.

There will also be a `CIRGenOpenCLRuntime.cpp` in the future as the
counterpart of `CGOpenCLRuntime.cpp` from OG CodeGen. But that's
basically a dedicated implementation file of a class `CGOpenCLRuntime`,
we should not bother it.