[CIR][Lowering] Add the concept of simple lowering and use it for unary fp2fp operations (#806)
This PR is the continuation and refinement of PR #434 which is
originally authored by @philnik777 . Does not update it in-place since I
don't have commit access to Nikolas' repo.

This PR basically just rebases #434 onto the latest `main`. I also
updated some naming used in the original PR to keep naming styles
consistent.

Co-authored-by: Nikolas Klauser <nikolasklauser@berlin.de>