[CIR][CodeGen][NFCI] Unify attribute list handling of func / call by `constructAttributeList` (#831)
Similar to #830 , this PR completes the `setCIRFunctionAttributes` part
with the call to `constructAttributeList` method, so that func op and
call op share the logic of handling these kinds of attributes, which is
the design of OG CodeGen.

It also includes other refactors. The function `constructAttributeList`
now use `mlir::NamedAttrList &` rather than immutable attribute
`mlir::DictionaryAttr &` as the inout result parameter, which benefits
the additive merging of attributes.