[CIR][CIRGen][NFC] Exceptions: add cleanups to cir.call
Now that the basic is working, start adding cleanups to be attached to
cir.call's instead. This is necessary in order to tie the pieces
(landing pads and cleanups) more properly, allowing multiple calls
inside cir.try op to be connected with the right cleanup.

This is the first piece of a series, tests coming next.